chore: retire shared/ shim and refresh top-level docs
Integration cleanup for unified-CLI refactor:

- Inline shared/pem_common.py into sn_confluent/core/pem.py and delete
  the shared/ directory (its docstring already telegraphed this).
- Update load_pem_files error message to point at `sn-confluent extract`
  instead of the retired `extract_pem.py` script.
- README.md: rewrite around the unified `sn-confluent` entry point and
  the sn_confluent/<subcommand>/ package layout.
- CONTRIBUTING.md: replace per-tool layout description with the new
  package structure; document core/ as the frozen shared API.
- pyproject.toml: drop `shared*` from package discovery.

Test count unchanged at 109; full suite green.
