+++ CONTRIBUTING.md
@@ -1,79 +1,71 @@
 # Contributing
 
-## Repo structure
+## Repo layout
 
 ```
-servicenow-confluent-tools/
-├── shared/               # Cross-cutting constants and utilities (stdlib only)
-│   └── pem_common.py
-├── extract-pem/          # Tool: PEM extraction from Java keystores
-├── cluster-link/         # Tool: Confluent Cloud cluster link creation
-├── mirror-topics/        # Tool: topic mirroring with interactive UI
-└── connect-replicator/   # Tool: Confluent Replicator deployment
+sn-confluent/
+├── pyproject.toml
+├── sn_confluent/
+│   ├── cli.py                # unified entry-point dispatcher
+│   ├── core/                 # shared API (frozen — see core/CONTRACT.md)
+│   │   ├── pem.py            #   PEM / Confluent CLI helpers
+│   │   └── config.py         #   link.conf loader + post-processing
+│   ├── extract/              # subcommand: PEM extraction
+│   ├── link/                 # subcommand: cluster link creation
+│   ├── mirror/               # subcommand: topic mirroring
+│   ├── replicate/            # subcommand: Replicator deployment
+│   └── setup/                # subcommand: end-to-end wizard
+└── README.md
 ```
 
-Each tool directory is independently runnable. `shared/` contains code used by two or more tools — it has no dependencies beyond the Python standard library.
+Each subcommand is a Python package under `sn_confluent/` with its own `main.py`, `README.md`, tests, and (if applicable) a `link.conf.example`. The unified CLI dispatcher in `sn_confluent/cli.py` lazy-imports each subcommand's `main(argv)` and forwards the remaining argv slice.
 
-## Adding a new tool
+## Adding a new subcommand
 
-Each tool lives in its own top-level directory:
+1. Create `sn_confluent/<name>/` with:
+   ```
+   __init__.py
+   main.py            # exports `main(argv: Optional[List[str]] = None) -> int`
+   README.md
+   tests/
+       __init__.py
+       test_<name>.py
+   ```
+2. Wire dispatch in `sn_confluent/cli.py` (`SUBCOMMAND_HELP` + `_load_subcommand`).
+3. If the subcommand needs orchestration in `sn-confluent setup`, add it to `sn_confluent/setup/wizard.py`.
+4. Update the table in the root `README.md`.
 
-```
-your-tool-name/
-├── README.md
-├── your_tool.py
-├── requirements.txt
-├── requirements-dev.txt
-└── tests/
-    └── test_your_tool.py
-```
+## Using `sn_confluent.core`
 
-- Include a `README.md` with setup, usage, and all CLI flags documented
-- Tests go in `tests/` and must pass with `pytest` run from the tool directory
-- Add the tool to the table in the root `README.md`
-
-## Using and extending `shared/`
-
-Tools import from `shared/` using a `sys.path` insert at the top of the script:
+`sn_confluent/core/` is the **frozen** shared API used by every subcommand. See `sn_confluent/core/CONTRACT.md` for the full contract.
 
 ```python
-import os, sys
-_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-if _ROOT not in sys.path:
-    sys.path.insert(0, _ROOT)
-from shared.pem_common import check_confluent_cli, load_pem_files, ...
+from sn_confluent.core.pem import (
+    SN_SOURCE_CLUSTERS, SN_BROKERS_PER_CLUSTER, CONFLUENT_INSTALL,
+    check_confluent_cli, check_auth, load_pem_files,
+)
+from sn_confluent.core.config import load_config, expand_link_config, add_per_cluster_link_names
 ```
 
-**What belongs in `shared/`:**
-- Constants used by 2+ tools (e.g. `SN_SOURCE_CLUSTERS`, `CONFLUENT_INSTALL`)
-- Utility functions called by 2+ tools with identical or near-identical behavior
-- stdlib-only code — no third-party dependencies
-
-**What stays tool-local:**
-- Business logic specific to one tool
-- Functions that differ meaningfully between tools
-- Anything requiring a third-party library
-
-When adding to `shared/`, update every tool that should use it and add a test in the relevant tool's test suite (shared code is tested through its callers, not standalone).
+If you need new behaviour from `core/`, propose the API change in a separate PR rather than amending `core/` mid-feature. The contract is what other subcommand maintainers rely on.
 
 ## Development setup
 
 ```bash
-cd <tool-directory>
-pip install -r requirements-dev.txt
-pytest
+pip install -e .[dev]
+pytest sn_confluent/
 ```
 
-No separate install is needed for `shared/` — it uses only the standard library.
+`pytest` from the repo root picks up every subcommand's test suite.
 
 ## Pull requests
 
-- One tool or fix per PR
-- All tests must pass before requesting review
-- Update the tool's `README.md` if flags or behavior change
+- One subcommand or fix per PR.
+- All tests must pass before requesting review.
+- Update the affected subcommand's `README.md` if flags or behaviour change.
 
 ## Secrets and credentials
 
-- Never commit PEM files, keystores, passwords, API keys, or live cluster IDs
-- Use `link.conf.example` as the pattern for config templates — real configs are gitignored
-- If you add a new tool with config, gitignore the live config and provide an `.example` file
+- Never commit PEM files, keystores, passwords, API keys, or live cluster IDs.
+- Use `link.conf.example` as the pattern for config templates — real configs are gitignored.
+- New subcommands with config files must provide an `.example` file and gitignore the live one.
