+++ README.md
@@ -1,35 +1,53 @@
-# ServiceNow — Confluent Tools
+# sn-confluent
 
-A collection of CLI tools for integrating ServiceNow's Kafka infrastructure with Confluent Cloud.
+A unified CLI for integrating ServiceNow's Kafka infrastructure with Confluent Cloud.
 
 > **Disclaimer:** This is personal work and is not supported, endorsed, or affiliated with ServiceNow or Confluent. Use at your own risk.
 
 ## Requirements
 
+- Python 3.9+
 - **Confluent CLI** — required for cluster link and topic mirroring operations. [Download here](https://docs.confluent.io/confluent-cli/current/install.html)
 
-## Tools
+## Install
 
-| Tool | Description |
+```bash
+pip install -e .
+```
+
+Installs the `sn-confluent` console entry point.
+
+## Subcommands
+
+| Subcommand | Description |
 |---|---|
-| [`extract-pem/`](extract-pem/) | Extract `ca.pem`, `client-cert.pem`, and `client-key.pem` from Java KeyStore / TrustStore files |
-| [`cluster-link/`](cluster-link/) | Create a Confluent Cloud Cluster Link from a source Kafka cluster using mTLS |
-| [`mirror-topics/`](mirror-topics/) | Mirror ServiceNow Kafka topics to Confluent Cloud across both DC cluster links with a checkbox UI |
-| [`connect-replicator/`](connect-replicator/) | Deploy a Confluent Replicator connector between ServiceNow Hermes and Confluent Cloud (bidirectional, interactive wizard) |
+| `sn-confluent extract`   | Extract `ca.pem`, `client-cert.pem`, and `client-key.pem` from PKCS12 keystores |
+| `sn-confluent link`      | Create a Confluent Cloud cluster link from a source Kafka cluster using mTLS |
+| `sn-confluent mirror`    | Mirror ServiceNow Kafka topics to Confluent Cloud across both DC cluster links with a checkbox UI |
+| `sn-confluent replicate` | Deploy a Confluent Replicator connector between ServiceNow Hermes and Confluent Cloud (bidirectional, interactive wizard) |
+| `sn-confluent setup`     | Guided end-to-end wizard that orchestrates the four steps above in one run |
+
+Run `sn-confluent <subcommand> --help` for command-specific options. Each subcommand also has its own README at `sn_confluent/<subcommand>/README.md`.
 
 ## Typical workflow
 
 ```bash
-# Step 1 — extract PEM files from JKS keystores
-cd extract-pem
-pip install -r requirements.txt
-python extract_pem.py --keystore ../path/to/keystore --truststore ../path/to/truststore --out-dir /tmp/pems
+# Step 1 — extract PEM files from PKCS12 keystores
+sn-confluent extract --keystore /path/to/keystore --truststore /path/to/truststore --out-dir /tmp/pems
 
 # Step 2 — create the cluster link
-cd ../cluster-link
-pip install -r requirements.txt
-cp link.conf.example link.conf   # fill in your cluster IDs
-python create_link.py --pem-dir /tmp/pems
+cp sn_confluent/link/link.conf.example sn_confluent/link/link.conf  # fill in your cluster IDs
+sn-confluent link --pem-dir /tmp/pems
+
+# Step 3 — mirror topics interactively
+sn-confluent mirror --pem-dir /tmp/pems
+
+# Step 4 — deploy Replicator (alternative to native cluster links)
+sn-confluent replicate --pem-dir /tmp/pems
 ```
 
-See each tool's README for full options.
+Or run all four with the guided wizard:
+
+```bash
+sn-confluent setup
+```
